started work on social media component
