+++ frontend/src/app/home/containers/Home.js
@@ -15,32 +15,9 @@ export default class Home extends React.Component {
   }
 
   componentDidMount(){
-    this.getWeather().then(() => this.getStocks()).then(() => this.getFeed());
-
+    this.getWeather().then(() => this.getStocks());
   }
 
-  getFeed() {
-    const jwt = localStorage.getItem("jwt");
-    const url = `api/twitter_proxy/feed`;
-
-    return fetch(url,
-      {
-        method: 'GET',
-        headers: {
-          'Accept': 'application/json',
-          'Content-Type': 'application/json',
-          'jwt': jwt
-        }
-      }
-    )
-    .then((res) => res.json())
-        .then((res) => {
-          this.setState({
-            feed: JSON.parse(res.feed_response)
-          });
-        }
-      );
-  }
 
   getWeather(){
     const url = 'https://api.openweathermap.org/data/2.5/weather?zip=14623,us&APPID=4c442afc1ade3bc91a9bb48fb1fd0e02&units=imperial';
